running by Alex

+++ data_scripts/get_data.py
@@ -34,7 +34,8 @@ def get_datasets():
     # get google global mobility
     url = "https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv"
     response = requests.get(url).content
-    os.remove("Global_Mobility_Report.csv")
+    if os.path.exists("Global_Mobility_Report.csv"):
+        os.remove("Global_Mobility_Report.csv")
     mobility_csv = open("Global_Mobility_Report.csv", 'wb')
     mobility_csv.write(response)
     mobility_csv.close()
@@ -45,16 +46,19 @@ def get_datasets():
     #get CDC hospitalization data
     url = "https://healthdata.gov/api/views/g62h-syeh/rows.csv?accessType=DOWNLOAD"
     response = requests.get(url).content
-    os.remove("COVID-19_Reported_Timeseries.csv")
-    hospital_csv = open("COVID-19_Reported_Timeseries.csv")
+    if os.path.exists("COVID-19_Reported_Timeseries.csv"):
+        os.remove("COVID-19_Reported_Timeseries.csv")
+    hospital_csv = open("COVID-19_Reported_Timeseries.csv",'wb')
     hospital_csv.write(response)
-    mobility_csv.close()
+    hospital_csv.close()
     print("CDC Hospitalization Data ... DONE")
+    quit()
 
     # get jhu casecount data
     url = "https://github.com/CSSEGISandData/COVID-19/raw/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_US.csv"
     response = requests.get(url).content
-    os.remove("time_series_covid19_confirmed_US.csv")
+    if os.path.exists("time_series_covid19_confirmed_US.csv"):
+        os.remove("time_series_covid19_confirmed_US.csv")
     jhu_confirmed_csv = open("time_series_covid19_confirmed_US.csv", 'wb')
     jhu_confirmed_csv.write(response)
     jhu_confirmed_csv.close()
@@ -63,7 +67,8 @@ def get_datasets():
     # get jhu deaths data
     url = "https://github.com/CSSEGISandData/COVID-19/raw/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_US.csv"
     response = requests.get(url).content
-    os.remove("time_series_covid19_deaths_US.csv")
+    if os.path.exists("time_series_covid19_deaths_US.csv"):
+        os.remove("time_series_covid19_deaths_US.csv")
     jhu_deaths_csv = open("time_series_covid19_deaths_US.csv", 'wb')
     jhu_deaths_csv.write(response)
     jhu_deaths_csv.close()
@@ -72,7 +77,8 @@ def get_datasets():
     # get pcr test data
     url = "https://healthdata.gov//api/views/j8mb-icvb/rows.csv?accessType=DOWNLOAD"
     response = requests.get(url).content
-    os.remove("COVID-19_PCR_Testing_Time_Series.csv")
+    if os.path.exists("COVID-19_PCR_Testing_Time_Series.csv"):
+        os.remove("COVID-19_PCR_Testing_Time_Series.csv")
     pcr_test_csv = open("COVID-19_PCR_Testing_Time_Series.csv", 'wb')
     pcr_test_csv.write(response)
     pcr_test_csv.close()
@@ -81,7 +87,8 @@ def get_datasets():
     # get excess death
     url = "https://data.cdc.gov/api/views/xkkf-xrst/rows.csv?accessType=DOWNLOAD&bom=true&format=true%20target="
     response = requests.get(url).content
-    os.remove("Excess_Deaths_COVID-19.csv")
+    if os.path.exists("Excess_Deaths_COVID-19.csv"):
+        os.remove("Excess_Deaths_COVID-19.csv")
     excess_death_csv = open("Excess_Deaths_COVID-19.csv", 'wb')
     excess_death_csv.write(response)
     excess_death_csv.close()
@@ -90,7 +97,8 @@ def get_datasets():
     # get vaccine data
     url = "https://github.com/govex/COVID-19/raw/master/data_tables/vaccine_data/us_data/time_series/vaccine_data_us_timeline.csv"
     response = requests.get(url).content
-    os.remove("vaccine_data_us_state_timeline.csv")
+    if os.path.exists("vaccine_data_us_state_timeline.csv"):
+        os.remove("vaccine_data_us_state_timeline.csv")
     vaccine_csv = open("vaccine_data_us_state_timeline.csv", 'wb')
     vaccine_csv.write(response)
     vaccine_csv.close()
